Added AP example and connection manager example to docs, updated docs and unified connection related examples
