+++ docs/starting_methods.rst.license
@@ -1,0 +1,3 @@
+SPDX-FileCopyrightText: 2024 Michał Pokusa
+
+SPDX-License-Identifier: MIT
