+++ examples/httpserver_simpletest_manual_ethernet.py
@@ -1,19 +1,21 @@
 # SPDX-FileCopyrightText: 2023 Tim C for Adafruit Industries
+#
 # SPDX-License-Identifier: MIT
 
 import board
 import digitalio
 
 from adafruit_wiznet5k.adafruit_wiznet5k import WIZNET5K
-import adafruit_wiznet5k.adafruit_wiznet5k_socket as socket
+from adafruit_wiznet5k import adafruit_wiznet5k_socket as socket
 from adafruit_httpserver import Server, Request, Response
 
-print("Wiznet5k HTTPServer Test")
 
 # For Adafruit Ethernet FeatherWing
 cs = digitalio.DigitalInOut(board.D10)
+
 # For Particle Ethernet FeatherWing
 # cs = digitalio.DigitalInOut(board.D5)
+
 spi_bus = board.SPI()
 
 # Initialize ethernet interface with DHCP
@@ -22,7 +24,6 @@ eth = WIZNET5K(spi_bus, cs)
 # Set the interface on the socket source
 socket.set_interface(eth)
 
-# Initialize the server
 server = Server(socket, "/static", debug=True)
